added logging

+++ src/midisds_send.c
@@ -7,17 +7,16 @@
 #include <sndfile.h>
 
 #include "midisds_common.h"
-#include "midisds_err.h"
 #include "midisds_rawmidi.h"
 #include "midisds_send.h"
 
 // Send an sds message to a device.
-int midisds_send_msg(const midi_t *midi, const midisds_message *msg) {
+ssize_t midisds_send_msg(const midi_t *midi, const midisds_message *msg) {
     midisds_header *hdr = msg->hdr;
     return midisds_send_header(midi, hdr);
 }
 
-int midisds_send_header(const midi_t *midi, midisds_header *hdr) {
+ssize_t midisds_send_header(const midi_t *midi, midisds_header *hdr) {
     size_t hdr_len = midisds_header_length();
     return midisds_send(midi, *hdr, hdr_len);
 }
