+++ src/midisds_send.h
@@ -1,6 +1,8 @@
 #ifndef MIDISDS_SEND_H
 #define MIDISDS_SEND_H
 
+#include <stdio.h>
+
 #include "midisds_common.h"
 #include "midisds_rawmidi.h"
 
@@ -20,7 +22,7 @@ typedef enum {
     RESPONSE_WAIT
 } response_t;
 
-int midisds_send_msg(const midi_t *midi, const midisds_message *msg);
-int midisds_send_header(const midi_t *midi, midisds_header *hdr);
+ssize_t midisds_send_msg(const midi_t *midi, const midisds_message *msg);
+ssize_t midisds_send_header(const midi_t *midi, midisds_header *hdr);
 
 #endif
